LayoutTests/imported/w3c/web-platform-tests/workers/semantics/run-a-worker/003-expected.txt

CONSOLE MESSAGE: Worker failed to load. Response is not 2xx

PASS worker
PASS shared
